inference/requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
transformers==4.40.0
torch==2.2.0
pydantic==1.10.11
python-dotenv==1.0.0
